vpn: fix multiple issues and split VPN managers
fix: #2956, #2946 avoid accessing DNS proxy details from the main context

fix: #2955 show IP protocol mismatch in a separate notification channel

fix: #2948 ensure the notification channel exists before posting notifications

fix: #2944 fix port rule checks in the upstream answer flow

refactor: split BraveVPNService into TunDnsManager, TunFirewallManager, and TunFlowManager
