+++ app/src/main/java/com/celzero/bravedns/service/DomainRulesManager.kt
@@ -27,6 +27,7 @@ import com.celzero.bravedns.RethinkDnsApplication.Companion.DEBUG
 import com.celzero.bravedns.database.CustomDomain
 import com.celzero.bravedns.database.CustomDomainRepository
 import com.celzero.bravedns.util.Constants
+import com.celzero.bravedns.util.Utilities.isAtleastR
 import com.celzero.firestack.backend.Backend
 import com.celzero.firestack.backend.RadixTree
 import org.koin.core.component.KoinComponent
@@ -222,6 +223,58 @@ object DomainRulesManager : KoinComponent {
         val res = Status.getStatus(status.toIntOrNull())
         if (DEBUG) Logger.vv(LOG_TAG_DNS, "domain rule for $key, res: $res")
         return res
+    }
+
+    // evaluates domain rules for the given domain(s) and uid.
+    // on Android R and above, only the first domain is considered, as the Go backend supplies it
+    // as the accurate domain. Returns the resolved Status along with the matched domain (trust
+    // takes precedence over block); NONE otherwise. Used by the tunnel managers.
+    fun getAggregatedDomainRule(domain: String?, uid: Int): Pair<Status, String?> {
+        if (domain.isNullOrEmpty()) {
+            return Pair(Status.NONE, "")
+        }
+
+        val domains = if (isAtleastR()) {
+            // on Android R and above, go will give the first domain as the accurate domain so
+            // no need to check further domains
+            val d = domain.lowercase(Locale.getDefault()).split(",").firstOrNull()
+            if (d.isNullOrEmpty()) return Pair(Status.NONE, "")
+            listOf(d)
+        } else {
+            domain.lowercase(Locale.getDefault()).split(",")
+        }
+
+        if (domains.isEmpty()) {
+            return Pair(Status.NONE, "")
+        }
+
+        var hasTrustedDomain = false
+        var trustedDomain = ""
+        var hasBlockedDomain = false
+        var blockedDomain = ""
+        for (d in domains) {
+            when (status(d, uid)) {
+                Status.TRUST -> {
+                    hasTrustedDomain = true
+                    trustedDomain = d
+                }
+
+                Status.BLOCK -> {
+                    hasBlockedDomain = true
+                    blockedDomain = d
+                }
+
+                else -> {
+                    // no-op
+                }
+            }
+        }
+
+        return when {
+            hasTrustedDomain -> Pair(Status.TRUST, trustedDomain)
+            hasBlockedDomain -> Pair(Status.BLOCK, blockedDomain)
+            else -> Pair(Status.NONE, "")
+        }
     }
 
     fun getProxyForDomain(uid: Int, domain: String): Pair<String, String> {
